# Add benchmark to measure scan performance on different setups