--- a/.docker/Dockerfile
+++ b/.docker/Dockerfile
@@ -1,5 +1,7 @@
 FROM golang:1.15-alpine
 
+RUN apk add --update --no-cache alpine-sdk
+
 WORKDIR /pgxscan
 
 COPY go.mod .
@@ -8,6 +10,4 @@ COPY go.sum .
 RUN go mod download
 COPY . ./
 
-RUN apk add --update --no-cache alpine-sdk
-
-CMD go test -v -race --tags=integration ./...
+CMD go test -v -race -bench=. -benchmem --tags=integration ./...