@randallmlough 2021-02-09
Adding benchmarks is a great addition. Although, I'm not sure if a benchmark that's dependent on a db connection is trustworthy. My goal of splitting the integration tests is to actually try and create a mock interface of pgx row/rows to test all the functionality without requiring a live db connection. Just haven't found the time to take a real stab at it yet