@randallmlough 2021-02-09
I'm going to merge this in since the comparison is pgx standard functionality and the libraries scan methods, but I do think it's important to not focus on the benchmarks too much here because to the acquiring of a db connection.