@pausan 2021-02-09
I agree. This is just to have some sort of a feeling on how are things doing, but as I said the benchmark is far from perfect.